Add files via upload
Both of these files are amended from the previous code for the Logistic Regression solver 'lbfgs' because it ran maximum iterations before concluding to results. 

I have replaced that with the 'liblinear' solver and also changed the limit of maximal iterations to 2000, although that does not affect the result once the 'liblinear' solver is deployed.